fix(templates): move binding modules to lib and drop lint suppressions
Restores the ratchet metrics the previous commit raised:
- cross-handler-imports: binding-sources/binding-catalog/build-binding-context/
  apply-source-fields move from handlers/docx to lib/template-binding (shared
  code belongs in lib; handler imports of lib are not cross-slice)
- lint-suppression-directives: fetchParties batches the per-role reads into one
  role-IN query (bounded by LIMITS.workspaceContactsCount) grouped in JS,
  removing the no-db-await-in-loop disable; the binding-catalog handler yields
  the static catalog through Result.ok instead of a require-yield disable
